SSM to enable S3 Bucket Encryption"
This script enables encryption on the specified S3 bucket using the AWS Key Management Service (KMS) customer master key provided. It makes use of the aws:executeAwsApi action to call the PutBucketEncryption API of the S3 service. The script also allows for an optional Automation Assume Role to be provided for cross-account access.